requirements.txt
boto3
botocore
pillow
